Add files via upload
Support Vector Machines (SVM) are a powerful and versatile class of supervised machine
learning algorithms used for classification and regression tasks. Here's some essential theory
about SVMs:
Data Set:
We download the Iris data set and use it for the present case. As we are performing the
practical on Google Colab, we upload the dataset as iris.csv